backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-generativeai
python-dotenv
